requirements.txt
alembic==1.15.2
astroid==3.3.10
bcrypt==4.3.0
blinker==1.9.0
cfgv==3.4.0
click==8.2.0
dill==0.4.0
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
identify==2.6.10
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
pylint==3.3.7
python-dotenv==1.1.0
PyYAML==6.0.2
SQLAlchemy==2.0.41
tomlkit==0.13.2
typing_extensions==4.13.2
virtualenv==20.31.2
Werkzeug==3.1.3
WTForms==3.2.1
